Simplify with __new__ on entity.
