+++ geoh5py/io/h5_reader.py
@@ -28,6 +28,7 @@ import numpy as np
 from ..data.float_data import FloatData
 from ..data.integer_data import IntegerData
 from ..shared import fetch_h5_handle
+from ..shared.concatenate import Concatenator
 from .utils import as_str_if_uuid, key_map, str2uuid, str_from_utf8_bytes
 
 
@@ -82,10 +83,8 @@ class H5Reader:
             if "PropertyGroups" in entity.keys():
                 property_groups = cls.fetch_property_groups(file, uid)
 
-            attributes["entity"]["concatenator"] = False
-
             if "Concatenated Data" in entity.keys():
-                attributes["entity"]["concatenator"] = True
+                attributes["entity"]["concatenation"] = Concatenator
 
             attributes["entity"]["on_file"] = True
