+++ geoh5py/objects/object_base.py
@@ -15,7 +15,6 @@
 #  You should have received a copy of the GNU Lesser General Public License
 #  along with geoh5py.  If not, see <https://www.gnu.org/licenses/>.
 
-# pylint: disable=W0613
 
 from __future__ import annotations
 
@@ -30,7 +29,6 @@ from ..data import CommentsData, Data
 from ..data.primitive_type_enum import PrimitiveTypeEnum
 from ..groups import PropertyGroup
 from ..shared import Entity
-from ..shared.concatenate import Concatenated
 from .object_type import ObjectType
 
 if TYPE_CHECKING:
@@ -46,19 +44,6 @@ class ObjectBase(Entity):
     _attribute_map.update(
         {"Last focus": "last_focus", "PropertyGroups": "property_groups"}
     )
-
-    def __new__(cls, *args, **kwargs):
-        if kwargs.get("concatenated", False):
-            for key, item in Concatenated.__dict__.items():
-                if "__" in key:
-                    continue
-
-                if "attribute_map" in key:
-                    cls._attribute_map.update(item)
-                else:
-                    setattr(cls, key, getattr(Concatenated, key))
-
-        return super().__new__(cls)
 
     def __init__(self, object_type: ObjectType, **kwargs):
         assert object_type is not None
